include signup and login button in greeting component
