+++ frontend/components/restaurant/restaurant_search_page_nav.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import styled from 'styled-components';
 import { Button, Label } from '../styles';
 import { fetchRestaurants } from '../../actions/restaurant_actions';
-import { UserATag } from "../styles";
+import Greeting from '../greeting/greeting';
 
 const Span = styled.span`
     margin-right: 10px;
@@ -55,18 +55,6 @@ const RestaurantSearchForm = styled.form`
     align-items: center; 
 `;
 
-const UserATagLogIn = styled(UserATag)`
-    margin-left: 20px;
-
-    &:hover {
-        background-color: #a71c1c;
-    }
-`;
-
-const UserATagSignUp = styled(UserATag)`
-    margin-left: 5px;
-`;
-
 const mapDispatchToProps = dispatch => {
     return {
         fetchRestaurants: (location, offset) => dispatch(fetchRestaurants(location, offset))
@@ -113,12 +101,7 @@ class RestaurantSearchPageNav extends React.Component {
                   </svg>
                 </SearchButton>
               </RestaurantSearchForm>
-              <UserATagLogIn href="login/" buttonColor="#d32323" buttonTextColor="#fff" border="1px solid #bd1f1f">
-                Log In
-              </UserATagLogIn>
-              <UserATagSignUp href="signup" buttonColor="#f5f5f5" buttonTextColor="#000" border="1px solid #f5f5f5">
-                Sign Up
-              </UserATagSignUp>
+              <Greeting />
             </RestaurantNavContainer>
           </RestaurantTopNavBar>;
     }
